Struct pandacoin::command_line_opts::CommandLineOpts

pub struct CommandLineOpts {
    pub key_path: String,
    pub password: Option<String>,
    pub genesis: bool,
    pub nocapture: bool,
}

This is the Clap options structure which stores all command line flags passed by the user to the application

Fields
key_path: String

Path to key-file

password: Option<String>genesis: boolnocapture: bool