Struct pandacoin::merkle_tree_manager::MerkleTree

pub struct MerkleTree { /* private fields */ }

Implementations

impl MerkleTree

An example of inner node indexes:

                  8
                /   \
               /     \
             4        12
            / \       / \
           /   \     /   \
          2    6    10    14
         /\    /\   /\    /\
        /  \  /  \ /  \  /  \
        1  3  5  7 9 11 13 15 <- we don't store these, they resolve to the tx hashes below
        |  |  |  | |  |  |  |
       T0 T1 T2 T3 T4 T5 T6 T7 <- leafs are transaction hashes

If the number of tx in a block is not exactly equal to some 2^n, we pad the tx set(leafs) with [0x0]s We refer to the upper/inner nodes of the tree as โnodesโ. For N leafs we need N-1 nodes, we will also store a [0x0] at the 0 position of the nodes array. This makes the indexes in the diagram above look the way they do and simplifies the logic/math of mapping those indexes into the nodes Vec and also gives us some valid data to return in case we want a merkle tree built from an empty set of Transactions.

pub fn new<'a, I>(iterator: I) -> Self
where I: Iterator<Item = &'a Transaction>,

pub fn new_from_hashes<'a, I>(iterator: I) -> Self
where I: Iterator<Item = &'a Sha256Hash>,

pub fn compute_node_hash_at_index( nodes: &Vec<Option<Sha256Hash>>, leafs: &Vec<Sha256Hash>, index: usize, ) -> Sha256Hash

pub fn get_nodes(&self) -> &[Option<Sha256Hash>]

pub fn get_root(&self) -> &Option<Sha256Hash>

pub fn make_sub_tree_for_index(&self, index: usize) -> MerkleTreeSubTree

pub fn validate_sub_tree(&self, sub_tree: MerkleTreeSubTree) -> bool

pub fn next_power_of_2(size: usize) -> usize