Trait pandacoin::block::RawBlock

pub trait RawBlock: Sync + Debug + Send {
    fn get_transactions(&self) -> &Vec<TransactionProto>;

    fn get_signature(&self) -> Signature { ... }

    fn get_hash(&self) -> &Sha256Hash { ... }

    fn get_block_fee(&self) -> u64 { ... }

    fn get_timestamp(&self) -> u64 { ... }

    fn get_creator(&self) -> PublicKey { ... }

    fn get_previous_block_hash(&self) -> Sha256Hash { ... }

    fn get_id(&self) -> u32 { ... }
}

This structure is a basic block, it should be 1-to-1 with the physical data which will be serialized and send over the wire and stored on disk. We provide a useless default implementation for the sake of making different mock RawBlocks easier. We would prefer to define only the interface here and the default implementation in mock_block.rs, but Rust does not allow this.

Required methods

fn get_transactions(&self) -> &Vec<TransactionProto>

Provided methods

fn get_signature(&self) -> Signature

fn get_hash(&self) -> &Sha256Hash

fn get_block_fee(&self) -> u64

fn get_timestamp(&self) -> u64

fn get_creator(&self) -> PublicKey

fn get_previous_block_hash(&self) -> Sha256Hash

fn get_id(&self) -> u32